refactor: rename router files and auth types file
About the routers,
The names "auth.ts" or "backoffice.ts" are worse for fuzzy finding then
"authRouter.ts" or "backofficeRouter.ts"

About the types, some of the types defined in that file are not that
related to auth or even api per say.

They are types related to the server / DB.
Also, with new incoming types like types for "product", this file would
start to have too many things or we would break the pattern by
processing or Prisma types in other places throughout the server...
IMO it's better to centralize these prisma / tRPC type processings in a
single file structure.