refactor: small adjustments
I believe most of these changes make the code more maintainable.
Be it by repeating a pattern already present in a lot of places in the
app (like the cn styling) or making the query param usage centralized
and less prone to bugs if they change in the future. Unfortunatelly
NextJS router is not typesafe... so a change in the query param can
break only at runtime.