feat(login|signup): change login & signup state management and routes
The state management is no longer a complex zustand machine
It's now a combination of query params and a very simple zustand state
specifically for storing signup information during the register flow

Also the BuyerForm and SellerForm components are now simpler, they don't
do API mutations, that is delegated to whomever is using them. This
makes much more sense, cus the usage for these components is completely
different in the Profile edit feature.